picom script updated to handle toggle vs always on behavior

+++ .bin/picom.sh
@@ -1,7 +1,13 @@
-#!/bin/zsh
+#!/bin/bash
 
-if pgrep -u $UID -x picom >/dev/null; then
-    killall -qw picom
+if [[ "$1" = "toggle" ]]; then
+    if pgrep -u $UID -x picom >/dev/null; then
+        killall -qw picom
+    else
+        picom --daemon --experimental-backends
+    fi
 else
-    picom --daemon --experimental-backends
+    if ! pgrep -u $UID -x picom >/dev/null; then
+        picom --daemon --experimental-backends
+    fi
 fi
